nbctf/nbctf-2024/qualifs/Pepe-temperature/infra/requirements.txt
Django>=4.2
paho-mqtt>=1.6.1
